Overhaul dotfiles so its compatible with Linux
Since getting a new razer blade gaming laptop, I have switched to linux
for personal development. As a result, I needed to add more Ubuntu
oriented configs into my scripts.

+++ zsh/install.sh
@@ -11,6 +11,7 @@ function install_deps() {
   logInfo "Installing Python and powerline"
   pip install --upgrade pip setuptools
   pip install powerline-status
+  npm install --global pure-prompt # For the "pure" prompt. See: https://github.com/sindresorhus/pure
 
   mkdir -p $HOME/bin/
 
@@ -20,8 +21,12 @@ function install_deps() {
   fi
 
   logInfo "Downloading and installing powerline appropriate fonts"
-  git clone https://github.com/powerline/fonts.git
-  cd fonts && ./install.sh && cd .. && rm -rf fonts
+  if uname | grep -q "Darwin"; then
+    git clone https://github.com/powerline/fonts.git
+    cd fonts && ./install.sh && cd .. && rm -rf fonts
+  else
+    sudo apt install fonts-powerline -y
+  fi
 
   logInfo "Setting the default shell to zsh. You maybe prompted for sudo access"
   chsh -s $(which zsh)
@@ -29,7 +34,7 @@ function install_deps() {
 
 function install_zim() {
   logInfo "Installing and configuring zim for a superior Zsh experience"
-  zsh && source $SCRIPT_DIR/install_zim.zsh
+  (exec $SCRIPT_DIR/install_zim.zsh)
 }
 
 function configure_autocompletion() {
@@ -73,11 +78,24 @@ function symlink_zimrc() {
   ln -s $SCRIPT_DIR/.zimrc $HOME/.zimrc
 }
 
+# For mac, 'rupa/z' tool is installed automatically when 'brew bundle' is installed.
+# As always, Linux requires extra work.
+function augment_shell_with_z_plugin_for_linux() {
+  if (( $(uname) == "Linux" )); then
+    logInfo "Downloading rupa/z for quick bouncing around into directories"
+    mkdir -p ~/shell-scripts
+    wget https://raw.githubusercontent.com/rupa/z/master/z.sh -O ~/shell-scripts/z.sh
+    echo ". ~/shell-scripts/z.sh" >> ~/.bashrc
+    echo ". ~/shell-scripts/z.sh" >> ~/.zshrc
+  fi
+}
+
 install_deps
 install_zim
 configure_autocompletion
 configure_autosourcer
 symlink_zimrc
+augment_shell_with_z_plugin_for_linux
 
 logInfo "Adding config to .zshrc"
 echo "ZSH_CUSTOM=${SCRIPT_DIR}/config" >> $HOME/.zshrc
